// packages/frontend/src/generated/types/CreateDispatchCompositeRequest.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { CreateDispatchRequest } from './CreateDispatchRequest.ts'
import type { DispatchItemCompositeRequest } from './DispatchItemCompositeRequest.ts'
import type { DispatchMeasurementCompositeRequest } from './DispatchMeasurementCompositeRequest.ts'

export type CreateDispatchCompositeRequest = CreateDispatchRequest & {
  /**
   * @type array
   */
  items: DispatchItemCompositeRequest[]
  /**
   * @type array,null
   */
  storageMeasurements?: DispatchMeasurementCompositeRequest[] | null
}