// packages/frontend/src/generated/zod/createDispatchCompositeRequestSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ToZod } from '../.kubb/ToZod.ts'
import type { CreateDispatchCompositeRequest } from '../types/CreateDispatchCompositeRequest.ts'
import * as z from 'zod'
import { createDispatchRequestSchema } from './createDispatchRequestSchema.ts'
import { dispatchItemCompositeRequestSchema } from './dispatchItemCompositeRequestSchema.ts'
import { dispatchMeasurementCompositeRequestSchema } from './dispatchMeasurementCompositeRequestSchema.ts'

export const createDispatchCompositeRequestSchema = z
  .lazy(() => createDispatchRequestSchema)
  .and(
    z.object({
      items: z.array(z.lazy(() => dispatchItemCompositeRequestSchema)),
      storageMeasurements: z
        .array(z.lazy(() => dispatchMeasurementCompositeRequestSchema))
        .nullish(),
    }),
  ) as unknown as ToZod<CreateDispatchCompositeRequest>