// packages/frontend/src/generated/zod/dispatchMeasurementCompositeRequestSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ToZod } from '../.kubb/ToZod.ts'
import type { DispatchMeasurementCompositeRequest } from '../types/DispatchMeasurementCompositeRequest.ts'
import * as z from 'zod'

export const dispatchMeasurementCompositeRequestSchema = z.object({
  afterDensity: z.string().nullish(),
  afterHeight: z.string().nullish(),
  afterMass: z.string(),
  afterVolume: z.string().nullish(),
  beforeDensity: z.string().nullish(),
  beforeHeight: z.string().nullish(),
  beforeMass: z.string(),
  beforeVolume: z.string().nullish(),
  storageId: z.string().uuid(),
}) as unknown as ToZod<DispatchMeasurementCompositeRequest>